sps-agent/requirements.txt
python-telegram-bot>=21.6
python-dotenv>=1.0.0
google-genai>=0.7.0
sqlalchemy>=2.0.40
pymysql>=1.1.1
openai>=1.30.0
ddgs>=8.0.0
